fantasy_backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.1.4
djangorestframework==3.15.2
fonttools==4.56.0
idna==3.10
kiwisolver==1.4.8
kombu==5.4.2
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
